Fix error when removing property on clipboard across devices

+++ src/utils/individual-tweaks.psm1
@@ -69,8 +69,14 @@ function Disable-ClipboardSyncAcrossDevice() {
     }
 
     Set-ItemProperty -Path "$PathToLMPoliciesSystem" -Name "AllowCrossDeviceClipboard" -Type DWord -Value 0
-    Remove-ItemProperty -Path "$PathToCUClipboard" -Name "CloudClipboardAutomaticUpload"
-    Remove-ItemProperty -Path "$PathToCUClipboard" -Name "EnableCloudClipboard"
+    If ((Get-Item "$PathToCUClipboard").Property -contains "CloudClipboardAutomaticUpload") {
+        Remove-ItemProperty -Path "$PathToCUClipboard" -Name "CloudClipboardAutomaticUpload"
+    }
+
+    If ((Get-Item "$PathToCUClipboard").Property -contains "EnableCloudClipboard") {
+        Remove-ItemProperty -Path "$PathToCUClipboard" -Name "EnableCloudClipboard"
+    }
+
 }
 
 function Enable-ClipboardSyncAcrossDevice() {
